requirements.txt
flask
python-dotenv
google-generativeai
pypdf
youtube-transcript-api
gunicorn
